mcp_flight_server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sse-starlette==2.1.3
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1

